requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.3.1+cpu
torchvision==0.18.1+cpu
sentence-transformers==3.0.1
langchain==0.2.14
langchain-chroma==0.1.3
langchain-community==0.2.12
langchain-huggingface==0.0.3
langchain-text-splitters==0.2.2
